+++ src/modules/home/components/Running/RunningStats.tsx
@@ -3,8 +3,7 @@ import { View, Text } from 'react-native';
 import Icon from '../../../../components/icon/Icon';
 import { TYPOGRAPHY } from '../../../../constants/typography';
 import { styles } from '../../styles/Running/RunningStats';
-import Geolocation from 'react-native-geolocation-service';
-import calculateDistance from '../../../../utils/calculateDistance';
+import useRunningTracker from '../../hooks/useRunningTracker';
 
 interface RunningStatsProps {
   distance?: number;
@@ -18,98 +17,11 @@ interface Location {
   longitude: number;
 }
 
-const RunningStats: React.FC<RunningStatsProps> = ({ 
-  distance = 0, 
-  time = 0, 
+const RunningStats: React.FC<RunningStatsProps> = ({
   isRunning = false,
   headerHeight = 68
 }) => {
-  const [currentTime, setCurrentTime] = useState(0);
-  const [currentDistance, setCurrentDistance] = useState(0);
-  const [timeComponents, setTimeComponents] = useState({ hours: '00', minutes: '00', seconds: '00' });
-  const intervalRef = useRef<NodeJS.Timeout | null>(null);
-  const locationWatchId = useRef<number | null>(null);
-  const lastLocation = useRef<Location | null>(null);
-
-  const formatTime = (totalSeconds: number) => {
-    const hours = Math.floor(totalSeconds / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-    
-    return {
-      hours: hours.toString().padStart(2, '0'),
-      minutes: minutes.toString().padStart(2, '0'),
-      seconds: seconds.toString().padStart(2, '0')
-    };
-  };
-
-  useEffect(() => {
-    if (isRunning) {
-      setCurrentTime(0);
-      setCurrentDistance(0);
-      lastLocation.current = null;
-
-      intervalRef.current = setInterval(() => {
-        setCurrentTime(prev => prev + 1);
-      }, 1000);
-
-      locationWatchId.current = Geolocation.watchPosition(
-        (position) => {
-          const newLocation: Location = {
-            latitude: position.coords.latitude,
-            longitude: position.coords.longitude
-          };
-
-          if (lastLocation.current) {
-            const distanceInMeters = calculateDistance(
-              lastLocation.current.latitude,
-              lastLocation.current.longitude,
-              newLocation.latitude,
-              newLocation.longitude
-            );
-            setCurrentDistance(prev => prev + (distanceInMeters / 1000));
-          }
-
-          lastLocation.current = newLocation;
-        },
-        (error) => {
-          console.log('위치 추적 에러:', error);
-        },
-        {
-          enableHighAccuracy: true,
-          distanceFilter: 5,
-          interval: 1000,
-        }
-      );
-    } else {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-        intervalRef.current = null;
-      }
-
-      if (locationWatchId.current) {
-        Geolocation.clearWatch(locationWatchId.current);
-        locationWatchId.current = null;
-      }
-
-      setCurrentTime(0);
-      setCurrentDistance(0);
-      lastLocation.current = null;
-    }
-
-    return () => {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-      }
-      if (locationWatchId.current) {
-        Geolocation.clearWatch(locationWatchId.current);
-      }
-    };
-  }, [isRunning]);
-
-  useEffect(() => {
-    setTimeComponents(formatTime(currentTime));
-  }, [currentTime]);
+  const { currentDistance, currentTime, timeComponents } = useRunningTracker(isRunning);
 
   const dynamicStyles = {
     ...styles.container,
